Change extract(epoch from renew_date) query to return f64 instead of i64

+++ binaries/geph5-broker/src/database/bandwidth.rs
@@ -3,7 +3,7 @@ use geph5_broker_protocol::BwConsumptionInfo;
 use super::POSTGRES;
 
 pub async fn bw_consumption(user_id: i32) -> anyhow::Result<Option<BwConsumptionInfo>> {
-    let pair: Option<(i32, i32, i64)> = sqlx::query_as("select (renew_mb - (mb_limit - mb_used)), renew_mb, extract(epoch from renew_date) from bw_usage natural join bw_limits where id = $1").bind(user_id).fetch_optional(&*POSTGRES).await?;
+    let pair: Option<(i32, i32, f64)> = sqlx::query_as("select (renew_mb - (mb_limit - mb_used)), renew_mb, extract(epoch from renew_date) from bw_usage natural join bw_limits where id = $1").bind(user_id).fetch_optional(&*POSTGRES).await?;
     Ok(pair.map(|pair| BwConsumptionInfo {
         mb_used: pair.0.max(0) as _,
         mb_limit: pair.1.max(0) as _,
